Move Simon board files to lib/simon for pure generic engine separation #173
