+++ package.json
@@ -51,7 +51,7 @@
 	},
 	"name": "simon",
 	"private": true,
-	"version": "0.72.0",
+	"version": "0.73.0",
 	"packageManager": "pnpm@10.33.2+sha512.a90faf6feeab71ad6c6e57f94e0fe1a12f5dcc22cd754db40ae9593eb6a3e0b6b12e3540218bb37ae083404b1f2ce6db2a4121e979829b4aff94b99f49da1cf8",
 	"type": "module",
 	"pnpm": {
