+++ src/lib/simon/board-config.ts
