+++ src/lib/simon/board-types.spec.ts
