+++ src/lib/simon/board-types.ts
