+++ src/lib/simon/types.ts
@@ -1,2 +1,2 @@
-export type { ButtonColor } from '$lib/game/board-types';
+export type { ButtonColor } from './board-types';
 export type SimonPhase = 'idle' | 'showing' | 'player_input' | 'round_complete' | 'gameover';
